로그아웃 기능구현, 작성자 저장, CRUD에서 로그인 체크

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -12,6 +12,7 @@ import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
 @Controller
 public class UsrArticleController {
 
@@ -20,7 +21,6 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/getArticle")
 	@ResponseBody
-	
 	public ResultData<Article> getArticle(int id) {
 
 		Article article = articleService.getArticleById(id);
@@ -35,44 +35,66 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-		public ResultData<Article> doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 하고 써");
+		}
 
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
-			
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
 		}
 
 		articleService.modifyArticle(id, title, body);
 
-		
 		article = articleService.getArticleById(id);
+
 		return ResultData.from("S-1", Ut.f("%d번 게시글을 수정했습니다", id), article);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 하고 써");
+		}
 
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
-			
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id), id);
 		}
 
 		articleService.deleteArticle(id);
 
-		
 		return ResultData.from("S-1", Ut.f("%d번 게시글을 삭제했습니다", id), id);
 	}
 
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession httpSession, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 하고 써");
+		}
 
 		if (Ut.isEmptyOrNull(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
@@ -81,18 +103,18 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
 
-		ResultData writeArticleRd = articleService.writeArticle(title, body);
+		ResultData writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
 
 		int id = (int) writeArticleRd.getData1();
 
 		Article article = articleService.getArticleById(id);
 
-		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), article);
+	
+		return ResultData.newData(writeArticleRd, article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
-	
 	public ResultData<List<Article>> getArticles() {
 		List<Article> articles = articleService.getArticles();
 		return ResultData.from("S-1", "Article List", articles);
